+++ node/header.go
@@ -1,9 +1,10 @@
 package node
 
 import (
-	"github.com/thrisp/wSCP/xandle"
-	"github.com/thrisp/wSCP/xandle/monitor"
-	"github.com/thrisp/wSCP/xandle/window"
+	"github.com/wSCP/xandle"
+	"github.com/wSCP/xandle/monitor"
+	"github.com/wSCP/xandle/tag"
+	"github.com/wSCP/xandle/window"
 )
 
 type Header interface {
@@ -14,18 +15,21 @@ type Header interface {
 	SetWindow(window.Window)
 	Monitor() monitor.Monitor
 	SetMonitor(monitor.Monitor)
+	Tag() tag.Tag
+	SetTag(tag.Tag)
 }
 
 type header struct {
 	x xandle.Xandle
 	w window.Window
 	m monitor.Monitor
+	t tag.Tag
 }
 
-var EmptyHeader header = header{nil, nil, nil}
+var EmptyHeader header = header{nil, nil, nil, nil}
 
-func NewHeader(x xandle.Xandle, w window.Window, m monitor.Monitor) Header {
-	return &header{x, w, m}
+func NewHeader(x xandle.Xandle, w window.Window, m monitor.Monitor, t tag.Tag) Header {
+	return &header{x, w, m, t}
 }
 
 func (h *header) CurrentHeader() Header {
@@ -54,4 +58,12 @@ func (h *header) Monitor() monitor.Monitor {
 
 func (h *header) SetMonitor(m monitor.Monitor) {
 	h.m = m
+}
+
+func (h *header) Tag() tag.Tag {
+	return h.t
+}
+
+func (h *header) SetTag(t tag.Tag) {
+	h.t = t
 }