+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ##Changelog
 
-### site62 0.0.0 (21.03.2016)
+### 0.0.1 (30.03.2016)
+
+- package imports rewrite
+- modules extraction and consolidation(windows is functional, monitors pending)
+- filesystem configuration rewrite
+
+### 0.0.0 (21.03.2016)
 
 - project initialized